requirements.txt
# Зависимости для конвертера документов

# PDF
PyPDF2>=3.0.0

# Word (DOCX)
python-docx>=1.0.0

# Excel
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Веб-интерфейс
Flask>=3.0.0
Werkzeug>=3.0.0

# OpenAI API
openai>=1.0.0

# HTTP requests (для Jay Flow API)
requests>=2.31.0

# HTTP клиент для прокси (для OpenAI)
httpx>=0.24.0

# База данных и аутентификация
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.3
Werkzeug>=3.0.0  # Для хеширования паролей

